messy commit, remove wait for GTM promise
The reason the tests are wrapped in `then()` blocks is the `ensureSetup`
promise which we planned to remove later in this PR so ive fast
forwarded that work to simplify fixing these specs (in case you havent
followed the commit the tests always return green)

The reason we dont need the promise is that we only need an array called
`dataLayer` and as soon as GTM instantiates it will process anything in
the array.

Note: `page` event specs are still disabled because they arent dont
match the implementation. category, label etc dont get sent to GTM. Next
steps are to work out if this is needed.